CustomerDraft to interface. #942
