+++ sphere-models/src/it/java/io/sphere/sdk/products/ProductCrudIntegrationTest.java
@@ -2,18 +2,18 @@ package io.sphere.sdk.products;
 
 import io.sphere.sdk.channels.Channel;
 import io.sphere.sdk.channels.ChannelDraft;
-import io.sphere.sdk.channels.ChannelDraftDsl;
 import io.sphere.sdk.channels.commands.ChannelCreateCommand;
+import io.sphere.sdk.client.SphereRequest;
 import io.sphere.sdk.products.commands.ProductCreateCommand;
 import io.sphere.sdk.products.commands.ProductDeleteCommand;
 import io.sphere.sdk.products.commands.ProductUpdateCommand;
-import io.sphere.sdk.products.commands.updateactions.*;
+import io.sphere.sdk.products.commands.updateactions.AddPrice;
+import io.sphere.sdk.products.commands.updateactions.RemovePrice;
 import io.sphere.sdk.products.queries.ProductQuery;
 import io.sphere.sdk.producttypes.ProductType;
 import io.sphere.sdk.producttypes.commands.ProductTypeCreateCommand;
 import io.sphere.sdk.producttypes.queries.ProductTypeQuery;
-import io.sphere.sdk.queries.*;
-import io.sphere.sdk.client.SphereRequest;
+import io.sphere.sdk.queries.PagedQueryResult;
 import io.sphere.sdk.suppliers.SimpleCottonTShirtProductDraftSupplier;
 import io.sphere.sdk.suppliers.TShirtProductTypeDraftSupplier;
 import io.sphere.sdk.test.IntegrationTest;
@@ -27,18 +27,19 @@ import java.util.List;
 import java.util.StringJoiner;
 import java.util.stream.IntStream;
 
+import static io.sphere.sdk.channels.ChannelFixtures.cleanUpChannelByKey;
 import static io.sphere.sdk.models.DefaultCurrencyUnits.EUR;
-import static io.sphere.sdk.products.ProductProjectionType.*;
-import static io.sphere.sdk.suppliers.TShirtProductTypeDraftSupplier.*;
+import static io.sphere.sdk.products.ProductFixtures.deleteProductsProductTypeAndProductDiscounts;
+import static io.sphere.sdk.products.ProductProjectionType.STAGED;
+import static io.sphere.sdk.suppliers.TShirtProductTypeDraftSupplier.Colors;
 import static io.sphere.sdk.suppliers.TShirtProductTypeDraftSupplier.Sizes;
-import static io.sphere.sdk.test.SphereTestUtils.*;
+import static io.sphere.sdk.test.SphereTestUtils.MASTER_VARIANT_ID;
+import static io.sphere.sdk.test.SphereTestUtils.en;
 import static io.sphere.sdk.utils.SphereInternalLogger.getLogger;
 import static java.util.Arrays.asList;
 import static java.util.Locale.ENGLISH;
 import static java.util.stream.Collectors.toList;
 import static org.assertj.core.api.Assertions.assertThat;
-import static io.sphere.sdk.channels.ChannelFixtures.*;
-import static io.sphere.sdk.products.ProductFixtures.*;
 
 public class ProductCrudIntegrationTest extends IntegrationTest {
     private static ProductType productType;
