+++ sphere-models/src/test/java/io/sphere/sdk/customers/CustomerDraftTest.java
@@ -11,26 +11,26 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 public class CustomerDraftTest {
 
-    private static final CustomerDraft CUSTOMER_DRAFT = CustomerDraft.of(CustomerName.ofFirstAndLastName("John", "Smith"), randomString(), "secret");
+    private static final CustomerDraftDsl CUSTOMER_DRAFT = CustomerDraftDsl.of(CustomerName.ofFirstAndLastName("John", "Smith"), randomString(), "secret");
     private static final List<Address> ADDRESSES = asList(Address.of(DE), Address.of(GB));
 
     @Test
     public void toStringDoesNotIncludeThePassword() throws Exception {
         final String secret = "123456";
-        final CustomerDraft draft = CustomerDraft.of(CustomerName.ofFirstAndLastName("hello", "world"), randomString(), secret);
+        final CustomerDraft draft = CustomerDraftDsl.of(CustomerName.ofFirstAndLastName("hello", "world"), randomString(), secret);
         assertThat(draft.toString()).doesNotContain(secret);
     }
 
     @Test
     public void checksAddressesForBillingAddress() throws Exception {
-        final CustomerDraft draft = CUSTOMER_DRAFT.withAddresses(ADDRESSES).withDefaultBillingAddress(1);
+        final CustomerDraftDsl draft = CUSTOMER_DRAFT.withAddresses(ADDRESSES).withDefaultBillingAddress(1);
         assertThatThrownBy(() -> draft.withDefaultBillingAddress(10))
                 .isInstanceOf(IllegalArgumentException.class);
     }
 
     @Test
     public void checksAddressesForShippingAddress() throws Exception {
-        final CustomerDraft draft = CUSTOMER_DRAFT.withAddresses(ADDRESSES).withDefaultShippingAddress(0);
+        final CustomerDraftDsl draft = CUSTOMER_DRAFT.withAddresses(ADDRESSES).withDefaultShippingAddress(0);
         assertThatThrownBy(() -> draft.withDefaultShippingAddress(10))
                 .isInstanceOf(IllegalArgumentException.class);
     }
